Put back ESM helpers in a folder where we can use `.js` (#12919)
