+++ .gitignore
@@ -26,24 +26,20 @@ package-lock.json
 /packages/babel-compat-data/build
 
 /packages/babel-runtime/helpers/*.js
-/packages/babel-runtime/helpers/*/*.js
-/packages/babel-runtime/helpers/*/*.mjs
-!/packages/babel-runtime/helpers/toArray/*
-!/packages/babel-runtime/helpers/iterableToArray/*
-!/packages/babel-runtime/helpers/temporalRef/*
-!/packages/babel-runtime/helpers/typeof/*
+!/packages/babel-runtime/helpers/toArray.js
+!/packages/babel-runtime/helpers/iterableToArray.js
+!/packages/babel-runtime/helpers/temporalRef.js
+!/packages/babel-runtime/helpers/typeof.js
 /packages/babel-runtime/helpers/esm/*.js
 !/packages/babel-runtime/helpers/esm/toArray.js
 !/packages/babel-runtime/helpers/esm/iterableToArray.js
 !/packages/babel-runtime/helpers/esm/temporalRef.js
 
 /packages/babel-runtime-corejs2/helpers/*.js
-/packages/babel-runtime-corejs2/helpers/*/*.js
-/packages/babel-runtime-corejs2/helpers/*/*.mjs
-!/packages/babel-runtime-corejs2/helpers/toArray/*
-!/packages/babel-runtime-corejs2/helpers/iterableToArray/*
-!/packages/babel-runtime-corejs2/helpers/temporalRef/*
-!/packages/babel-runtime-corejs2/helpers/typeof/*
+!/packages/babel-runtime-corejs2/helpers/toArray.js
+!/packages/babel-runtime-corejs2/helpers/iterableToArray.js
+!/packages/babel-runtime-corejs2/helpers/temporalRef.js
+!/packages/babel-runtime-corejs2/helpers/typeof.js
 /packages/babel-runtime-corejs2/helpers/esm/*.js
 !/packages/babel-runtime-corejs2/helpers/esm/toArray.js
 !/packages/babel-runtime-corejs2/helpers/esm/iterableToArray.js
@@ -52,8 +48,6 @@ package-lock.json
 !/packages/babel-runtime-corejs2/core-js/map.js
 
 /packages/babel-runtime-corejs3/helpers/*.js
-/packages/babel-runtime-corejs3/helpers/*/*.js
-/packages/babel-runtime-corejs3/helpers/*/*.mjs
 /packages/babel-runtime-corejs3/helpers/esm/*.js
 /packages/babel-runtime-corejs3/core-js/**/*.js
 /packages/babel-runtime-corejs3/core-js-stable/**/*.js
