--- packages/babel-runtime-corejs2/helpers/typeof/_index.mjs
@@ -1,17 +1,0 @@
-import _Symbol from "@babel/runtime-corejs2/core-js/symbol";
-import _Symbol$iterator from "@babel/runtime-corejs2/core-js/symbol/iterator";
-export default function _typeof(obj) {
-  "@babel/helpers - typeof";
-
-  if (typeof _Symbol === "function" && typeof _Symbol$iterator === "symbol") {
-    _typeof = function _typeof(obj) {
-      return typeof obj;
-    };
-  } else {
-    _typeof = function _typeof(obj) {
-      return obj && typeof _Symbol === "function" && obj.constructor === _Symbol && obj !== _Symbol.prototype ? "symbol" : typeof obj;
-    };
-  }
-
-  return _typeof(obj);
-}
