+++ test/runtime-integration/expected-cjs-13.2.txt
@@ -3,7 +3,7 @@ typeof objectWithoutProperties: function
 typeof objectWithoutProperties.default: function
 obj: { b: 2, [Symbol(Symbol.toStringTag)]: 5 }
 ================= require - esm =====================
-Error: Cannot use import statement outside a module
+Error: Must use import to load ES Module: <ROOT>/packages/babel-runtime/helpers/esm/toPrimitive.js
 =============== require - corejs ====================
 typeof Set: function
 arr: 1,2,3
