+++ test/runtime-integration/expected-esm-13.2.txt
@@ -13,10 +13,6 @@ typeof objectWithoutProperties.default: function
 obj: { b: 2, [Symbol(Symbol.toStringTag)]: 5 }
 ================= require - esm =====================
 Error: Must use import to load ES Module: <ROOT>/packages/babel-runtime/helpers/esm/toPrimitive.js
-require() of ES modules is not supported.
-require() of <ROOT>/packages/babel-runtime/helpers/esm/toPrimitive.js from <ROOT>/test/runtime-integration/src/require-esm.cjs is an ES module file as it is a .js file whose nearest parent package.json contains "type": "module" which defines all .js files in that package scope as ES modules.
-Instead rename toPrimitive.js to end in .cjs, change the requiring code to use import(), or remove "type": "module" from <ROOT>/packages/babel-runtime/helpers/esm/package.json.
-
 =============== require - corejs ====================
 typeof Set: function
 arr: 1,2,3
